+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Users can be created both locally and in Auth0
 - Users can be associated with multiple organisations
 - Allow roles to be assigned to users
+- Users are welcomed and able to create their new password to access the service